Create hooks again
